pyproject.toml
[project]
name = "gemini-live-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "google-genai>=1.73.1",
    "pydantic-settings>=2.13.1",
    "websockets>=16.0",
]
